% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial02 package}
\alias{RnaSeqTutorial02}
\alias{RnaSeqTutorial02-package}
\alias{RnaSeqTutorial02 package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorials package
}
\section{Tutorials}{

This is the second in a series of tutorials
\itemize{
\item\code{02_exploratory_data_analysis} perform the exploratory data analysis of a public RNA-Seq dataset
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("02_exploratory_data_analysis", package = "RnaSeqTutorial02")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
